requirements.txt
Jinja2==3.1.4
Werkzeug==3.0.4
Flask==3.0.3
# CleanDoc - Dependencias
# ========================

# Framework web
flask==3.0.3

# Manipulación de archivos DOCX
python-docx==1.1.2

# Servidor WSGI para producción
gunicorn==22.0.0

# Parser XML (requerido por python-docx)
lxml==5.3.0

# Variables de entorno
python-dotenv==1.0.1

# Utilidades
werkzeug==3.0.4
